debug

+++ bert/modeling.py
@@ -167,7 +167,7 @@ class BertModel(object):
     # Add by Winfred
     if input_stroke_ids is not None:
       input_stroke_shape = get_shape_list(input_stroke_ids, expected_rank=2)
-      stroke_length = input_stroke_shape[1] / seq_length
+      stroke_length = int(input_stroke_shape[1] / seq_length)
       input_stroke_ids = tf.reshape(input_stroke_ids,
           shape=[batch_size, seq_length, stroke_length])
     # End
